requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
anaconda==0.0.1.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.5.0
certifi==2021.10.8
charset-normalizer==3.4.0
checksumdir==1.2.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dataclasses-json==0.6.7
distlib==0.3.7
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
faiss-cpu==1.9.0
filelock==3.12.2
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.28.5
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==2.0.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.3.2
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-google-genai==2.0.1
langchain-text-splitters==0.3.0
langsmith==0.1.136
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.23.0
matplotlib==3.5.1
mdurl==0.1.2
mealpy==3.0.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.4
numpy==1.26.4
oauthlib==3.2.2
opfunu==1.0.4
orjson==3.10.9
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
platformdirs==3.10.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyHTML==1.3.2
pyparsing==3.0.6
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
speaklater==1.3
SQLAlchemy==1.4.45
streamlit==1.39.0
tenacity==9.0.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.24.2
visitor==0.1.3
watchdog==5.0.3
Werkzeug==3.0.1
WTForms==3.0.1
yarl==1.15.5
